feat-#57: change struct type.

+++ test/clickhousev2/v2.42.0/test_clickhousev2_crud.go
@@ -36,7 +36,7 @@ var (
 	createTable = `CREATE TABLE IF NOT EXISTS users (
 		id String, 
 		name String,
-		age UInt8
+		age Int32
 	) ENGINE = MergeTree() ORDER BY id`
 
 	insert = `INSERT INTO users (id, name, age) VALUES (?, ?, ?)`
